feat: update db verification for validation
