+++ .talismanrc
@@ -135,7 +135,7 @@ fileignoreconfig:
 - filename: server/tests/data/opcoEp.ts
   checksum: 8d1f421390c62fce249bf9efa7bb0688c6c9aa97375dd83d45f28c7d2246bf42
 - filename: server/tests/unit/deca.test.ts
-  checksum: e10531aed92c66bf3a44ba55e94b841fff1a75d0b4ee9160e644e0df81053595
+  checksum: e3f6b8fe70f0793ee367971538658938ddfc0ba62e281e1e3bbde6f7aaac2347
 - filename: ui/.env
   checksum: 36cfa5591cf1dcf75cc7fb3be6980cc26a393786f0afffdcca92573c448b3148
 - filename: ui/.env.local.example
