+++ server/src/modules/jobs/deca/hydrate-organization-person-deca.ts
@@ -1,12 +1,11 @@
 import { pMapIterable } from "p-map";
 import { DOCUMENT_TYPES } from "shared/constants/documents";
-import { getSirenFromSiret } from "shared/helpers/common";
 import { IDeca } from "shared/models/deca.model/deca.model";
 
 import parentLogger from "@/common/logger";
 
 import { getDbCollection } from "../../../common/utils/mongodbUtils";
-import { updateOrganisationData } from "../../actions/organisations.actions";
+import { updateOrganisationAndPerson } from "../../actions/organisations.actions";
 
 const logger = parentLogger.child({ module: "job:hydrate:from_deca" });
 
@@ -16,42 +15,12 @@ async function runDoc(docDeca: IDeca) {
 
   if (!courriel || !siret) return;
 
-  const siren = getSirenFromSiret(siret);
-
   const countA = (courriel.match(/@/g) || []).length;
   if (countA > 1) return; // bad data multiple email
 
-  const domain = courriel.split("@")[1].toLowerCase();
-  if (!domain) return;
-
   try {
-    const organisation = await updateOrganisationData({
-      siren,
-      sirets: [siret],
-      email_domains: [domain],
-      source: DOCUMENT_TYPES.DECA,
-    });
-
-    await getDbCollection("persons").updateOne(
-      {
-        email: courriel,
-      },
-      {
-        $addToSet: {
-          ...(organisation && { organisations: organisation._id.toString() }),
-          sirets: siret,
-        },
-        $setOnInsert: {
-          email: courriel,
-        },
-      },
-      {
-        upsert: true,
-      }
-    );
+    await updateOrganisationAndPerson(siret, courriel, DOCUMENT_TYPES.DECA);
   } catch (error) {
-    // console.log(courriel, domains);
-    console.log(error.errInfo.details.schemaRulesNotSatisfied[0].propertiesNotSatisfied[0].details);
     console.log(error);
   }
 }
